Moved Prometheus registry from middleware to Celery
