+++ server/constants.py
@@ -1,0 +1,1 @@
+APP_NAME = "Dhruva"
