+++ server/middleware/prometheus_metrics_middleware.py
@@ -19,7 +19,6 @@ class PrometheusMetricsMiddleware(BaseHTTPMiddleware):
         custom_labels: Optional[Dict[str, Union[str, Callable]]] = None,
     ):
         super().__init__(app)
-        self.registry = CollectorRegistry()
         self.app_name = app_name.lower()
         self.custom_labels = custom_labels
 
@@ -38,73 +37,8 @@ class PrometheusMetricsMiddleware(BaseHTTPMiddleware):
             *self._get_custom_labels_values(request),
         ]
 
-        self.request_count.labels(*labels).inc()
-        end = time.perf_counter()
-        self.request_duration_seconds.labels(*labels).observe(end - begin)
-
-        push_metrics.apply_async((self.registry,), queue="metrics_log")
-
+        push_metrics.apply_async((labels, time.perf_counter() - begin,), queue="metrics_log")
         return response
-
-    @property
-    def request_count(self):
-        unit = "requests_total"
-        metric_name = f"{self.app_name}_{unit}"
-        key = f"{unit}_counter"
-
-        try:
-            metric = self.__getattribute__(key)
-        except Exception:
-            metric = Counter(
-                metric_name,
-                "Total HTTP requests",
-                registry=self.registry,
-                labelnames=(
-                    "method",
-                    "path",
-                    "status_code",
-                    "app_name",
-                    "request_id",
-                    *self._get_custom_labels_keys(),
-                ),
-            )
-
-            self.__setattr__(key, metric)
-
-        return metric
-
-    @property
-    def request_duration_seconds(self):
-        unit = "request_duration_seconds"
-        metric_name = f"{self.app_name}_{unit}"
-        key = f"{unit}_histogram"
-
-        try:
-            metric = self.__getattribute__(key)
-        except Exception:
-            metric = Histogram(
-                metric_name,
-                "Request duration seconds",
-                registry=self.registry,
-                labelnames=(
-                    "method",
-                    "path",
-                    "status_code",
-                    "app_name",
-                    "request_id",
-                    *self._get_custom_labels_keys(),
-                ),
-            )
-
-            self.__setattr__(key, metric)
-
-        return metric
-
-    def _get_custom_labels_keys(self):
-        if self.custom_labels is None:
-            return []
-
-        return list(self.custom_labels.keys())
 
     def _get_custom_labels_values(self, request: Request):
         if self.custom_labels is None:
